refactor towards better
