+++ src/paths/index.js
@@ -1,4 +1,5 @@
 import * as debug from './debug'
+import * as login from './login'
 
 /**
  * GET collection => Retrieve all resources in a collection
@@ -7,6 +8,7 @@ import * as debug from './debug'
  */
 export const gets = [
   ...debug.gets,
+  login.get,
   {
     path: '/',
     handler (req, res, next) {
